Last

+++ app/src/main/java/com/example/gtm/data/entities/custom/QuestionNewPost.kt
@@ -4,7 +4,7 @@ import com.example.gtm.data.entities.ui.Image
 
 data class QuestionNewPost(
     val questionId: Long? = null,
-    val rate: Long? = null,
+    val rate: Double? = null,
     val description: String? = null,
     val images: ArrayList<Image>? = null
 )
