+++ app/src/main/java/com/example/gtm/data/entities/remote/QuestionPost.kt
@@ -3,7 +3,7 @@ package com.example.gtm.data.entities.remote
 
 data class QuestionPost(
     val questionId: Long? = null,
-    val rate: Long? = null,
+    val rate: Double? = null,
     val description: Any? = null,
     val images: List<ImagePath?>? = null
 )
